retrived the data succefully

+++ dataI/src/main/java/com/example/datai/controller/UserController.java
@@ -1,6 +1,7 @@
 package com.example.datai.controller;
 
 import com.example.datai.dto.ProductDto;
+import com.example.datai.repository.UserRepo;
 import com.example.datai.service.impl.UserServiceImpl;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -18,7 +19,7 @@ public class UserController {
     private UserServiceImpl userService;
 
     @GetMapping(path = "{/id}")
-    public List<ProductDto> findAllProductCreatedByUser(@PathVariable("id") int id){
+    public List<ProductDto> findAllProductCreatedByUser(@PathVariable("id") long id){
         return userService.findAllProductByCreatedByUser(id);
     }
 }
