+++ dataI/src/main/java/com/example/datai/service/UserService.java
@@ -5,5 +5,7 @@ import com.example.datai.dto.ProductDto;
 import java.util.List;
 
 public interface UserService {
-    List<ProductDto> findAllProductByCreatedByUser(int id);
+
+    List<ProductDto> findAllProductByCreatedByUser(long id);
+
 }
